Implement all game rounds and evaluate in game.py
A few changes are made. check_player_action() should deal with the case
raised amount is not supported by the owned stack for a player. A new
method handle_betting_round() is refactored from pre_flop_round() as it
is a segment reused by all game rounds. The rest rounds are implemented
to simulate the full work flow.